Bug Fix: Fix Registrator access to Consul Client
When attempting to connect the Registrator container with the Consul client instance
that was local to the machine, I was see an error that it was unable to connect to
port 53 on the docker bridge network for the consul container.

In the consul client configuration, port 53 was being forward to port 8600, but this
only needs to be done on the Consul server to git unrestricted access to DNS since
port 53 is a restricted port. After changing to expose port 53 to host port 53,
everything is working as expected.

Also updated the scripts to track the host instance resources.
Disk, Memory, CPU.

CPU still needs to be updated to include the load on each core. Currently we
are only checking the total host cpu usage.